+++ credential-exchange-format/CHANGELOG.md
@@ -17,6 +17,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Changed
 
 - **BREAKING**: Changed `integration_hash` to `integrity_hash` in `FileCredential`. (#87)
+- **BREAKING**: Renamed `ty` to `type` in serialized representations of `Credential::Unknown`. (#125)
 
 ### Fixed