+++ credential-exchange-format/CHANGELOG.md
@@ -7,7 +7,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
-## [0.4.0] - 2026-06-08
+## [0.4.0] - 2026-06-09
 
 ### Added